Updated NK_small Testing
Updated NK_Nsmall.py to store all data in a pickle, as well as output
summary information to an excel sheet.

This was a fair amount of work, and involved completely rewriting the
package, essentially.

A similar approach should be done to preserve the outputs of sequence
tools.

Note: Results from initial NK_testing (messy) are still on desktop and
have not been synced